Убирает цикличность модулей
